Update CurrencyItems and NPCs
In NPCs, it looks like a string field for Gender was added, containing M
or F.

In CurrencyItems, a new bool field was added. It seems to be false for
every row except what looks like Gold in the latest row, so IsGold seems
a good enough guess for now.